+++ Thunderstore/CHANGELOG.md
@@ -2,7 +2,9 @@
 ## 1.3.0
 
 - Added On Hit Bomb Attack config option
+- Added Bomb Density config option
 - Checked the spelling of config titles and descriptions
+- Adjusted config in various ways to better support Gale (mod manager)
 - Fixed language override not working?
 
 ## 1.2.1 - *SotS 1.3.5 Fix*